refact : Car move 로직 Movable 객체로 추출
