+++ src/main/kotlin/racingcar/domain/Car.kt
@@ -1,29 +1,10 @@
 package racingcar.domain
 
-import java.lang.IllegalArgumentException
-
-class Car {
+class Car(private val movable: Movable) {
     private var location = 0
-    private val BASE_NUMBER_TO_MOVE = 4
-    private val RANDOM_MINIMUM = 0
-    private val RANDOM_MAXIMUM = 9
 
-    fun isMoveOrStop(): Boolean {
-        val randomNum = generateRandomNum()
-        val conditionToMove = conditionToMove(randomNum)
-        if (conditionToMove) location++
-        return conditionToMove
-    }
-
-    private fun conditionToMove(randomNum: Int) = randomNum >= BASE_NUMBER_TO_MOVE
-
-    fun generateRandomNum(): Int {
-
-        val randomValue: Int = (RANDOM_MINIMUM..RANDOM_MAXIMUM).random()
-        if (randomValue < RANDOM_MINIMUM || randomValue > RANDOM_MAXIMUM) {
-            throw IllegalArgumentException()
-        }
-        return randomValue
+    fun isMoveOrStop() {
+        if (movable.move()) location++
     }
 
     fun getLocation(): Int {
