+++ src/main/kotlin/racingcar/domain/Movable.kt
@@ -1,0 +1,5 @@
+package racingcar.domain
+
+interface Movable {
+    fun move() : Boolean
+}
