+++ src/main/kotlin/racingcar/domain/RandomMoveStrategy.kt
@@ -1,0 +1,28 @@
+package racingcar.domain
+
+import racingcar.utils.RandomGenerator
+
+class RandomMoveStrategy : Movable {
+
+    override fun move(): Boolean {
+        val randomNum = generateRandomNum()
+        return conditionToMove(randomNum)
+    }
+
+    private fun conditionToMove(randomNum: Int) = randomNum >= BASE_NUMBER_TO_MOVE
+
+    private fun generateRandomNum(): Int {
+
+        val randomValue: Int = RandomGenerator.generateRandomNum(RANDOM_MINIMUM, RANDOM_MAXIMUM)
+        if (randomValue < RANDOM_MINIMUM || randomValue > RANDOM_MAXIMUM) {
+            throw IllegalArgumentException("랜덤값은 0,9 사이의 값이어야 합니다.")
+        }
+        return randomValue
+    }
+
+    companion object {
+        private val BASE_NUMBER_TO_MOVE = 4
+        private val RANDOM_MINIMUM = 0
+        private val RANDOM_MAXIMUM = 9
+    }
+}
