+++ src/main/kotlin/racingcar/utils/RandomGenerator.kt
@@ -1,0 +1,11 @@
+package racingcar.utils
+
+import racingcar.domain.RandomMoveStrategy
+
+class RandomGenerator {
+    companion object {
+        fun generateRandomNum(begin: Int, end: Int) :Int {
+            return (begin..end).random()
+        }
+    }
+}
